requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file ./requirements.txt requirements.txt
#
aioauth-client==0.12.0
aiohttp==3.1.2
async-timeout==2.0.1      # via aiohttp
attrs==17.4.0             # via aiohttp, nicfit.py
certifi==2018.1.18        # via requests
chardet==3.0.4            # via aiohttp, requests
deprecation==2.0.1        # via nicfit.py
facebook-sdk==2.0.0
idna-ssl==1.0.1           # via aiohttp
idna==2.6                 # via idna-ssl, requests, yarl
multidict==4.1.0          # via aiohttp, yarl
nicfit.py==0.8
packaging==17.1           # via deprecation
pillow==5.1.0
pylast==2.2.0
pyparsing==2.2.0          # via packaging
requests==2.18.4
six==1.11.0               # via packaging, pylast
urllib3==1.22             # via requests
yarl==1.1.1               # via aiohttp
